requirements.txt
PyYAML==6.0
nltk==3.8.1
requests==2.30.0